// apps/desktop/src/ipc/generated/TaskState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TaskState = "queued" | "starting" | "running" | "blocked" | "paused" | "reported" | "takenOver" | "landing" | "readyToLand" | "landed" | "done" | "rejected" | "stopped" | "failed";